refactor: extract `formatSeconds` from `DateUtils`
